Setup /schools API endpoint database connection
